[Config] Match input and output vars to Tup's optional-extra shape
This adds support for specifying regular inputs and outputs in addition
to extra ones. A bit clunky for the toolchain implementations, which
now have to spell out the separate `extra_inputs` and `extra_outputs`
fields, but this will come in very handy for both Clang and libraries
in general. Surprisingly, this even *removes* source code lines?!

Part of P0299, funded by Ember2528 and Arandui.