Restrict C++ modules to a single input
The Clang implementation needs to pass a command-substituted path for
its `std` and `std.compat` modules, which needs to bypass Tup's input
file handling. Restricting the API to a single file keeps that
implementation much saner, and *probably* won't be an issue anyway?

Part of P0299, funded by Ember2528 and Arandui.